// src/Api/developer/dto/authorize.dto.ts
export class DeveloperAuthorizeApiDto {
  site: string;
  website: string;
  callback: string;
  email: string;
  tel: string;
  mobile: string;
  description: string;
  logo: string;
  title: string;
  postalcode: string;
  status: boolean;
  type: number;
  user: string;
  ip: any;
  apikey: string;
  id: string;
}